src/training-scripts/py/cifar10/requirements.txt
numpy==1.15.1
Pillow==8.4.0
tensorflow==2.7.0
